+++ src/point.cpp
@@ -1,4 +1,4 @@
-#include "point.h"
+#include "point.hpp"
 
 namespace snake {
